consoleアプリケーションとライブラリにプロジェクトを分割して書き直し
