+++ packages/Yasunori.NET/Yasunori.sln
@@ -1,9 +1,11 @@
-
+﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.0.31903.59
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Yasunori", "Yasunori.csproj", "{493E1711-CFE6-410E-BC46-86139BD0FDAF}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Yasunori.Net", "lib\Yasunori.Net.csproj", "{969A4E3E-ACF3-4276-A861-460B8C86004F}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Yasunori", "cli\Yasunori.csproj", "{D9975762-EEEC-461D-BF39-53D37D7A2DC3}"
 EndProject
 Global
 	GlobalSection(SolutionProperties) = preSolution
@@ -18,17 +20,29 @@ Global
 		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{493E1711-CFE6-410E-BC46-86139BD0FDAF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{493E1711-CFE6-410E-BC46-86139BD0FDAF}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{493E1711-CFE6-410E-BC46-86139BD0FDAF}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{493E1711-CFE6-410E-BC46-86139BD0FDAF}.Debug|x64.Build.0 = Debug|Any CPU
-		{493E1711-CFE6-410E-BC46-86139BD0FDAF}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{493E1711-CFE6-410E-BC46-86139BD0FDAF}.Debug|x86.Build.0 = Debug|Any CPU
-		{493E1711-CFE6-410E-BC46-86139BD0FDAF}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{493E1711-CFE6-410E-BC46-86139BD0FDAF}.Release|Any CPU.Build.0 = Release|Any CPU
-		{493E1711-CFE6-410E-BC46-86139BD0FDAF}.Release|x64.ActiveCfg = Release|Any CPU
-		{493E1711-CFE6-410E-BC46-86139BD0FDAF}.Release|x64.Build.0 = Release|Any CPU
-		{493E1711-CFE6-410E-BC46-86139BD0FDAF}.Release|x86.ActiveCfg = Release|Any CPU
-		{493E1711-CFE6-410E-BC46-86139BD0FDAF}.Release|x86.Build.0 = Release|Any CPU
+		{969A4E3E-ACF3-4276-A861-460B8C86004F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{969A4E3E-ACF3-4276-A861-460B8C86004F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{969A4E3E-ACF3-4276-A861-460B8C86004F}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{969A4E3E-ACF3-4276-A861-460B8C86004F}.Debug|x64.Build.0 = Debug|Any CPU
+		{969A4E3E-ACF3-4276-A861-460B8C86004F}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{969A4E3E-ACF3-4276-A861-460B8C86004F}.Debug|x86.Build.0 = Debug|Any CPU
+		{969A4E3E-ACF3-4276-A861-460B8C86004F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{969A4E3E-ACF3-4276-A861-460B8C86004F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{969A4E3E-ACF3-4276-A861-460B8C86004F}.Release|x64.ActiveCfg = Release|Any CPU
+		{969A4E3E-ACF3-4276-A861-460B8C86004F}.Release|x64.Build.0 = Release|Any CPU
+		{969A4E3E-ACF3-4276-A861-460B8C86004F}.Release|x86.ActiveCfg = Release|Any CPU
+		{969A4E3E-ACF3-4276-A861-460B8C86004F}.Release|x86.Build.0 = Release|Any CPU
+		{D9975762-EEEC-461D-BF39-53D37D7A2DC3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D9975762-EEEC-461D-BF39-53D37D7A2DC3}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D9975762-EEEC-461D-BF39-53D37D7A2DC3}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{D9975762-EEEC-461D-BF39-53D37D7A2DC3}.Debug|x64.Build.0 = Debug|Any CPU
+		{D9975762-EEEC-461D-BF39-53D37D7A2DC3}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{D9975762-EEEC-461D-BF39-53D37D7A2DC3}.Debug|x86.Build.0 = Debug|Any CPU
+		{D9975762-EEEC-461D-BF39-53D37D7A2DC3}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D9975762-EEEC-461D-BF39-53D37D7A2DC3}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D9975762-EEEC-461D-BF39-53D37D7A2DC3}.Release|x64.ActiveCfg = Release|Any CPU
+		{D9975762-EEEC-461D-BF39-53D37D7A2DC3}.Release|x64.Build.0 = Release|Any CPU
+		{D9975762-EEEC-461D-BF39-53D37D7A2DC3}.Release|x86.ActiveCfg = Release|Any CPU
+		{D9975762-EEEC-461D-BF39-53D37D7A2DC3}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
