+++ packages/Yasunori.NET/lib/Yasunori.cs
@@ -1,6 +1,6 @@
-namespace Yasunori;
+namespace Yasunori.Net;
 
-class Yasunori
+public class YasunoriData
 {
     public uint Id { get; set; }
     public string Title { get; set; } = string.Empty;
@@ -9,4 +9,9 @@ class Yasunori
     public string Senpan { get; set; } = string.Empty;
     public string Content { get; set; } = string.Empty;
     public string Meta { get ; set; } = string.Empty;
+
+    public override string ToString()
+    {
+        return $"{Id,3:d} {Date} {Title}";
+    }
 }
